+++ packages/list-plugins/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @keystonejs/list-plugins
 
+## 7.0.2
+
+### Patch Changes
+
+- [`0bf9e0615`](https://github.com/keystonejs/keystone/commit/0bf9e06153c52196b32c4e71e96b469cfe94f649) [#3160](https://github.com/keystonejs/keystone/pull/3160) Thanks [@timleslie](https://github.com/timleslie)! - Updated to use `context.executeGraphQL` for all server-side GraphQL operations.
+
+- Updated dependencies [[`4884ce609`](https://github.com/keystonejs/keystone/commit/4884ce6094b3c9ec203c702a5de97b983bd14176)]:
+  - @keystonejs/fields@13.0.1
+
 ## 7.0.1
 
 ### Patch Changes